@@ -4,33 +4,49 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
+    "build:prod": "ng build --configuration production",
     "build": "ng build",
+    "start:prod": "ng serve --configuration production",
     "watch": "ng build --watch --configuration development",
     "test": "ng test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^19.2.14",
-    "@angular/cdk": "^19.2.19",
-    "@angular/common": "^19.2.0",
-    "@angular/compiler": "^19.2.0",
-    "@angular/core": "^19.2.0",
-    "@angular/forms": "^19.2.0",
-    "@angular/material": "^19.2.19",
-    "@angular/platform-browser": "^19.2.0",
-    "@angular/platform-browser-dynamic": "^19.2.0",
-    "@angular/router": "^19.2.0",
-    "angularx-qrcode": "^19.0.0",
+    "@angular/animations": "^20.3.3",
+    "@angular/cdk": "^20.2.7",
+    "@angular/common": "^20.3.3",
+    "@angular/compiler": "^20.3.3",
+    "@angular/core": "^20.3.3",
+    "@angular/forms": "^20.3.3",
+    "@angular/material": "^20.2.7",
+    "@angular/platform-browser": "^20.3.3",
+    "@angular/platform-browser-dynamic": "^20.3.3",
+    "@angular/router": "^20.3.3",
+    "@angular/service-worker": "^20.3.3",
+    "@microsoft/signalr": "^9.0.6",
+    "@ng-icons/core": "^32.2.0",
+    "@ng-icons/huge-icons": "^32.4.0",
+    "@ng-icons/iconsax": "^32.2.0",
+    "@ng-icons/solar-icons": "^32.4.0",
+    "angularx-qrcode": "^20.0.0",
+    "chart.js": "^4.5.0",
     "dayjs": "^1.11.13",
+    "html-to-image": "^1.11.13",
+    "html2canvas": "^1.4.1",
+    "material-icons": "^1.13.14",
+    "ng2-charts": "^8.0.0",
+    "ngx-cookie-service": "^20.1.0",
     "qrcode": "^1.5.4",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.2.13",
-    "@angular/cli": "^19.2.13",
-    "@angular/compiler-cli": "^19.2.0",
+    "@angular/build": "^20.3.4",
+    "@angular/cli": "^20.3.4",
+    "@angular/compiler-cli": "^20.3.3",
+    "@angular/language-server": "^20.2.2",
+    "@angular/language-service": "^20.3.3",
     "@tailwindcss/postcss": "^4.1.12",
     "@types/jasmine": "~5.1.0",
     "@types/qrcode": "^1.5.5",
@@ -41,8 +57,9 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "postcss": "^8.5.6",
+    "postcss": "^8.5.3",
+    "prettier": "^3.6.2",
     "tailwindcss": "^4.1.12",
-    "typescript": "~5.7.2"
+    "typescript": "~5.9.3"
   }
 }
