@@ -20,7 +20,7 @@ export class BusinessGuard implements CanActivate {
       map((isAuthenticated) => {
         console.log(isAuthenticated);
         if (isAuthenticated) {
-          const t = this.authService.getCurrentUserSnapshot();
+          const t = this.authService.currentUserSnapshot;
           console.log(t);
           return true;
         } else {
