@@ -1,23 +1,18 @@
-import {
-  HttpEvent,
-  HttpHandler,
-  HttpInterceptor,
-  HttpRequest,
-} from '@angular/common/http';
+import { HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/common/http';
 import { Injectable } from '@angular/core';
+import { environment } from '@env/environment';
 import { Observable } from 'rxjs';
 
 @Injectable()
 export class CredentialsInterceptor implements HttpInterceptor {
-  intercept(
-    request: HttpRequest<unknown>,
-    next: HttpHandler
-  ): Observable<HttpEvent<unknown>> {
-			
-    const clonedRequest = request.clone({
-      withCredentials: true,
-    });
+  intercept(request: HttpRequest<unknown>, next: HttpHandler): Observable<HttpEvent<unknown>> {
+    if (request.url.startsWith(environment.apiUrl)) {
+      const clonedRequest = request.clone({
+        withCredentials: true,
+      });
+      return next.handle(clonedRequest);
+    }
 
-    return next.handle(clonedRequest);
+    return next.handle(request);
   }
 }
