@@ -3,8 +3,8 @@ import { Booking } from './Booking';
 export interface Client {
   id: string;
   fullName: string;
-  phoneNumber: string;
   email: string;
-  bookings: Booking[];
+  phoneNumber?: string;
+  bookings?: Booking[];
 	avatarUrl?: string;
 }
