@@ -1,7 +1,9 @@
 import { Injectable } from '@angular/core';
 import { objectToFormData } from '@app/shared/utils/object-to-form';
 import dayjs from 'dayjs';
-import { Observable } from 'rxjs';
+import { BehaviorSubject, filter, first, Observable, tap } from 'rxjs';
+import { BusinessSetupItem } from '../DTOs/business-setup-state.response';
+import { ClientsSummaryResponse } from '../DTOs/clients-summary-response';
 import { Business, TimeSlot } from '../models/Business';
 import { BusinessStatement } from '../models/BusinessStatement';
 import { CreateBusinessDTO } from '../schemas/create-business.dto';
@@ -11,20 +13,63 @@ import { ApiService } from './api.service';
 
 @Injectable({ providedIn: 'root' })
 export class BusinessService {
+  private businessSession$$ = new BehaviorSubject<Business | null>(null);
+  public businessSession$ = this.businessSession$$.asObservable();
+
   constructor(private readonly apiService: ApiService) {}
+
+  public loadBusinessSession(): Observable<Business> {
+    if (this.businessSession$$.getValue()) {
+      return this.businessSession$.pipe(
+        filter((business): business is Business => business !== null), // Garante que não é nulo
+        first(),
+      );
+    }
+
+    return this.getBusinessSession().pipe(
+      tap((business) => {
+        this.businessSession$$.next(business);
+      }),
+    );
+  }
+
+  public getBusinessSetupState(): Observable<BusinessSetupItem[]> {
+    return this.apiService.get('/business/setup-state');
+  }
+
+  public delete(): Observable<any> {
+    return this.apiService.delete('/business');
+  }
+
+  public updatePassword(newPassword: string): Observable<{ message: string }> {
+    return this.apiService.put('/business/password', { newPassword });
+  }
+
+  public removePin(): Observable<void> {
+    return this.apiService.delete('/business/password');
+  }
 
   public getAll(): Observable<Business[]> {
     return this.apiService.get<Business[]>('/business/all');
+  }
+
+  public avaibleHandle(handle: string): Observable<boolean> {
+    return this.apiService.get(`/business/avaible-handle/${handle}`);
   }
 
   public create(createBusinessDTO: CreateBusinessDTO): Observable<Business> {
     return this.apiService.post<Business>('/business', createBusinessDTO);
   }
 
-  public select(businessId: string): Observable<boolean> {
+  public select(businessId: string, password?: string): Observable<boolean> {
     return this.apiService.post('/auth/business', {
       businessId,
+      ...(password && { password }),
     });
+  }
+
+  public openOrClose(isOpen: boolean): Observable<{ status: boolean }> {
+    return this.apiService.put('/business/status', { isOpen });
   }
 
   public createTimeSlot(timeSlot: CreateTimeSlotDTO): Observable<TimeSlot> {
@@ -57,5 +102,9 @@ export class BusinessService {
 
   public getTimeSlots(): Observable<TimeSlot[]> {
     return this.apiService.get('/business/timeSlot');
+  }
+
+  public getClients(): Observable<ClientsSummaryResponse> {
+    return this.apiService.get('/business/clients');
   }
 }
