@@ -1,4 +1,8 @@
 export const environment = {
   production: true,
-  apiUrl: 'https://39b7f1aaf2a4.ngrok-free.app'
+  apiUrl: 'https://buy-document-aspect-burlington.trycloudflare.com/api',
+  mercadoPagoPublicKey: 'TEST-5d068378-41a9-4662-902f-cd922e8335e9',
+  // apiUrl: 'https://q5frp03ij64b.share.zrok.io/api',
+  // apiUrl: 'https://10.220.0.8:7001/api',
+  // apiUrl: 'http://10.220.0.8:5055/api',
 };
