@@ -1,4 +1,8 @@
 export const environment = {
-	production: false,
-	apiUrl: 'http://10.220.0.8:5055'
+  production: false,
+  apiUrl: 'http://10.220.0.8:5055/api',
+  mercadoPagoPublicKey: "TEST-5d068378-41a9-4662-902f-cd922e8335e9"
 };
+
+// apiUrl: 'https://buy-document-aspect-burlington.trycloudflare.com/api',
+// apiUrl: 'https://q5frp03ij64b.share.zrok.io/api',
